style dir restructure to colocation. migrate dweb post
