+++ app/components/index.ts
@@ -1,5 +1,5 @@
-export { default as Button } from "./button";
+export { default as Button } from "./button/button";
 export { default as Header } from "./header";
 export { default as Links } from "./links";
-export { default as Notification } from "./notification";
+export { default as Notification } from "./notification/notification";
 export { default as Root } from "./root";
