--- app/styles/index.module.css
@@ -1,11 +1,0 @@
-.column {
-  flex-direction: column;
-  flex-basis: 45%;
-  gap: var(--mantine-spacing-sm);
-}
-
-.column p,
-.column a {
-  font-weight: 700;
-  display: block;
-}
