+++ app/util.ts
@@ -1,3 +1,5 @@
+import { Mdx } from "types/modules";
+
 export function cls(...args: (string | undefined)[]) {
   return args.join(" ").trim();
 }
@@ -8,3 +10,14 @@ export const status = {
   ok: "ok",
 } as const;
 export type Status = (typeof status)[keyof typeof status];
+
+/**
+ * Save as variable since glob uses this when generating keys
+ */
+export const blogPath = "./blog";
+
+export function getPosts() {
+  return import.meta.glob<Mdx>("./blog/*.mdx", {
+    eager: true,
+  });
+}
