+++ vite.config.ts
@@ -9,6 +9,7 @@ export default defineConfig({
   plugins: [
     mdx({
       remarkPlugins: [remarkFrontmatter, remarkMdxFrontmatter],
+      providerImportSource: "@mdx-js/react",
     }),
     remix({
       ignoredRouteFiles: ["**/*.css"],
